Rename 'etsy_env' module to 'env'
