+++ etsy/__init__.py
@@ -1,5 +1,5 @@
 from _v2 import EtsyV2 as Etsy
-from etsy_env import EtsyEnvSandbox, EtsyEnvProduction
+from env import EtsyEnvSandbox, EtsyEnvProduction
 
 
 __version__ = '0.3.1'
@@ -7,4 +7,3 @@ __author__ = 'Dan McKinley'
 __copyright__ = 'Copyright 2010, Etsy Inc.'
 __license__ = 'GPL v3'
 __email__ = 'dan@etsy.com'
-
