+++ etsy/_v2.py
@@ -1,6 +1,6 @@
 import urllib
 from _core import API, missing
-from etsy_env import EtsyEnvSandbox, EtsyEnvProduction
+from env import EtsyEnvSandbox, EtsyEnvProduction
 
 try:
     from urlparse import parse_qsl
@@ -11,7 +11,7 @@ class EtsyV2(API):
     api_version = 'v2'
 
     def __init__(self, api_key='', key_file=None, method_cache=missing, 
-                 etsy_env=EtsyEnvSandbox(), log=None, etsy_oauth_client=None):
+                 etsy_env=EtsyEnvSandbox, log=None, etsy_oauth_client=None):
         self.api_url = etsy_env.api_url
         self.etsy_oauth_client = None
